clear project
