+++ index.html
@@ -3,7 +3,7 @@
 
 <head>
   <meta charset="UTF-8" />
-  <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+  <link rel="icon" type="image/svg+xml" href="./src/assets/plane-icon_yellow.png" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Vivu Airline</title>
 </head>
