+++ src/component/Home/FormSearch/FormSearch.css
@@ -1,0 +1,35 @@
+.buyForm{
+    background-color: rgb(255, 255, 255,0.33);
+    box-shadow: 0px 10px 32px #006885;
+    backdrop-filter: blur(3px);
+    padding: 0px 30px 20px 30px;
+    border-radius: 20px;
+
+}
+.radio{
+    padding-top: 20px;
+}
+.icon-search{
+    margin-right: 5px;
+      padding-top: 2px;
+      height: 20px;
+      padding-bottom: 0;
+      color: #006885;
+ }
+.label-children{
+        font-size: 12px !important;
+        color: grey;
+}
+.btn-search{
+    font-size: 20px;
+    font-family: 500;
+    color: white;
+    padding: 0px 20px 0px 20px;
+    font-size: 18px;
+    align-items: center;
+    font-weight: 600;
+    background-color: #006885;
+    &:hover{
+        color: black !important;
+    }
+}
