+++ src/component/SelectFlight/InfoAndStep/InfoAndstep.css
@@ -1,0 +1,8 @@
+.infor-select{ 
+    padding-top: 10px;
+    padding-left: 17%; 
+}
+.icon-selcet{
+    padding: 20px;
+    size: 2rem;
+}
