+++ src/component/SelectFlight/InfoFly/InfoFly.css
@@ -1,0 +1,72 @@
+
+.imgBusinesswhite{
+    border: 1px solid rgb(175, 137, 3);
+    background-color: rgb(175, 137, 3);
+    border-top-left-radius: 10px;
+    border-top-right-radius: 10px;
+    margin-left:5px ;
+}
+.imgskyboss{
+    border: 1px solid rgb(218, 33, 40);
+    background-color: rgb(218, 33, 40);
+    border-top-left-radius: 10px;
+    border-top-right-radius: 10px;
+    margin-left:5px ;
+
+}
+.imgVeluxe{
+    border: 1px solid rgb(249, 165, 26);
+    background-color: rgb(249, 165, 26);
+    border-top-left-radius: 10px;
+    border-top-right-radius: 10px;
+    margin-left:5px ;
+}
+.imgEco{
+    border: 1px solid rgb(106, 183, 46);
+    background-color: rgb(106, 183, 46);
+    border-top-left-radius: 10px;
+    border-top-right-radius: 10px;
+    margin-left:5px ;
+}
+.symbol-airline{
+    height: 100px;
+    background-color: #006885;
+    border-top-left-radius: 10px;
+    border-bottom-left-radius: 10px;
+}
+.price-airline{
+    height: 100px;
+    background-color: white;
+    margin-left:5px ;
+}
+.price{
+    font-size: 25px;
+    color: black;
+    font-weight: 600;
+    display: flex;
+    justify-content: center;
+    padding-top: 20px;
+}
+.price-vnd{
+    display: flex;
+    justify-content: center;
+    font-size: 20px;
+    color: grey;
+    font-weight: 600;
+}
+.code-airline{
+    display: flex;
+    justify-content: center;
+    font-size: 16px;
+    padding-top: 10px;
+    color: white;
+
+}
+.clock-airline{
+    font-size: 25px;
+    color: white;
+    font-weight: 600;
+    display: flex;
+    justify-content: center;
+
+}
