+++ src/component/SelectFlight/SelectInfoFly/SelectInfoFly.css
@@ -1,0 +1,24 @@
+.infor-user-select{
+    margin-top: 5%;
+    background-color:white;
+    box-shadow: 0px 10px 32px #006885;
+    backdrop-filter: blur(3px);
+    border-radius: 10px;
+}
+
+.title-select{
+    border-top-left-radius: 10px;
+    border-top-right-radius: 10px;
+    background-color: #006885;
+    height: 50px;
+    padding: 10px;
+
+}
+.title-infor{
+ margin: 20px 30px 20px 30px;
+ background-color: #F1F1F1;
+ border-radius: 10px;
+ height: 50px;
+ padding-top: 7px;
+
+}
