+++ src/component/SelectFlightInfor/CheckInfoFly/CheckInfoFly.css
@@ -1,0 +1,44 @@
+.select-fight-infor{
+    background-color: white;
+    height: auto;
+    width: 100%;
+    margin-top: 10px;
+    border-radius: 10px;
+    padding: 10px;
+}
+.date-fight{
+    font-size: 16px;
+    font-weight: 500;
+
+}
+.location-start{
+    font-size: 25px;
+    color: black;
+    font-weight: 600;
+    display: flex;
+    justify-content: center;
+    padding-top: 20px;
+}
+.time-start{
+    display: flex;
+    justify-content: center;
+    font-size: 18px;
+    color: grey;
+    font-weight: 600;
+}
+.flight-time{
+    padding-top: 20px;
+    display: flex;
+    justify-content: center;
+    font-size: 16px;
+    color: grey;
+    font-weight: 600;
+}
+.code-flight{
+    padding-top: 10px;
+    display: flex;
+    justify-content: center;
+    font-size: 16px;
+    color: black;
+    font-weight: 700;
+}
